+++ src/Certes.Cli/AzureClientFactory.cs
@@ -1,6 +1,0 @@
-﻿using Microsoft.Azure.Management.ResourceManager.Fluent.Core;
-
-namespace Certes.Cli
-{
-    internal delegate T AzureClientFactory<T>(RestClient restClient);
-}