+++ src/Certes.Cli/Commands/AzureOptions.cs
@@ -1,9 +1,0 @@
-﻿using Certes.Cli.Settings;
-
-namespace Certes.Cli.Commands
-{
-    internal class AzureOptions : AzureSettings
-    {
-        public string ResourceGroup { get; set; }
-    }
-}