+++ src/Certes.Cli/Commands/CommandGroup.cs
@@ -26,12 +26,6 @@
             Help = Strings.HelpCommandCertificate,
         };
 
-        public static readonly CommandGroup Azure = new CommandGroup
-        {
-            Command = "az",
-            Help = Strings.HelpCommandAzure,
-        };
-
         public string Command { get; private set; }
         public string Help { get; private set; }