+++ src/Certes.Cli/Program.cs
@@ -2,9 +2,6 @@
 using System.Reflection;
 using System.Threading.Tasks;
 using Autofac;
-using Microsoft.Azure.Management.AppService.Fluent;
-using Microsoft.Azure.Management.Dns.Fluent;
-using Microsoft.Azure.Management.ResourceManager.Fluent;
 using NLog;
 using NLog.Config;
 using NLog.Targets;
@@ -31,9 +28,6 @@ namespace Certes.Cli
             builder.RegisterType<CliCore>();
             builder.RegisterType<CliCoreSpectre>();
             builder.RegisterType<AcmeContext>().As<IAcmeContext>();
-            builder.RegisterType<DnsManagementClient>().As<IDnsManagementClient>();
-            builder.RegisterType<ResourceManagementClient>().As<IResourceManagementClient>();
-            builder.RegisterType<WebSiteManagementClient>().As<IWebSiteManagementClient>();
 
             return builder.Build();
         }